Update management_module.v to better integrate with execution engine
Rewrote the management module so it now has separate always blocks for handling the operational state FSM, individual FSMs for the shutdown state and startup state, and for handling the TPM response code output. Added in a response code input from the execution engine and a response code output. Removed all startup procedures performed by cryptographic sub-modules and added an output of startup type to be sent to the execution engine.